[overseer/cycle-26] Fix POINTS_SCALE env var truncating decimal values
Category: bugfix
Trigger: Phase 2b: parseInt("0.1") = 0 which is then filtered out, so
  setting POINTS_SCALE=0.1,0.5,1,2,3,5,8 via env would silently drop 0.1
  and 0.5 and produce an empty-ish scale

Changed parseInt → parseFloat in config.server.ts POINTS_SCALE parsing.
Added test case verifying decimal scale values round-trip correctly.

+++ src/lib/config.server.ts
@@ -83,7 +83,7 @@ function parseEnvToConfig(env: Env): Partial<AhaSMTConfig> {
     }
     if (env.POINTS_SCALE) {
       config.points.scale = env.POINTS_SCALE
-        .split(",").map((s) => parseInt(s.trim())).filter((n) => !isNaN(n));
+        .split(",").map((s) => parseFloat(s.trim())).filter((n) => !isNaN(n));
     }
     if (env.POINTS_DEFAULT_PER_DAY !== undefined) {
       config.points.defaultPerDay = env.POINTS_DEFAULT_PER_DAY;
